Adds page-level nav links based on headings
A list of links is generated from the h2 and h3 headings on each page and added to the top under the header. The original idea was to add them as expanding secitons in the sidebar menu nav, but in most cases users wouldn't see them until after scrolling down the page.

Also added a styling detail to highlight the current page in the sidebar nav.