Add rather complicated CSS graphic between menu and page nav
A CSS-only solution using ::before and ::after on existing elements to add a line between the current page in the menu, and the page-level navigation links for that page. Other than a rather liberal use of calc(), it's actaully not all that messy, except that in the case of the first few items in the menu list, the lines have to be turned the other way (downward) and this is done manually with page 1Ds one by one.