Update full theme to Alpha 14
When updating the theme for this project, long with the `variables.scss` file, there are aslo customizations in the `source/cs/application.js file` that need to be carried over for the custom menu behavior (visible at widths greater than a tablet). In that js file, see lines 183-197 for menuSetup(), and 288-290 for the winow resize event.