Update CONTRIBUTING.md
Just made a few small changes. Mainly to "ways of getting involved" and "first-timers." I also added our GitHub tutorial/article to the "Submit your contribution" section. I read through everything else more closely and I think all our bases are covered, including the possibility of contributions being edited or rejected.  

Question: For the link you have for Join Us (similar to Documentation and Forum)... is that just a place holder? Or once changes are merged into master will it direct you to the actual site? Or does it just take you to that section of the GH repository? Speaking of, the original link for the roadmap just went to the documentation so I updated it (hopefully correctly) and also created one for the GitHub tutorial.