Update publication-opportunities.md
further updates to pre-publication resources

though a table format seems to work well for the funding portion; it does not work as well necessarily for the rest of the resources outlined, so that part is inconsistent. Not sure if that's a problem.

Furthermore, the funding resources are listed in alphabetical order using the acronyms/shortened forms, because they are more commonly used for reference.

That column is called "shortform" in order to accommodate the fact that not all of the abbreviated names are acronyms, technically.