restore v0 of cli-commands screenshot
The previously deleted v0 of the screenshot cli-commands has been restored and a new id has been created for it. The shortcode on the v0 of the quire docs has been updated to reflect the change of id. Now a different version of the cli-commands screenshot shows in each version of quire docs